requirements.txt
openai
pdfplumber
pydantic
python-dotenv
requests
openai==1.0.0
reportlab
streamlit
pandas
